w3c-reports/12092014/efl/runned-tests/html/semantics/scripting-1/the-script-element/script-language-type-expected.txt


FAIL A script with empty type should run assert_equals: expected true but got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/scripting-1/the-script-element/script-language-type.html:16:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/scripting-1/the-script-element/script-language-type.html:15:5)
